git-svn-id: https://svn.code.sf.net/p/cspfj/code/trunk/cspfj@506 b3bf8218-cbbe-40ea-ae7b-cf512e877a40

+++ src/main/java/cspfj/filter/AC3.java
@@ -40,7 +40,7 @@ public final class AC3 implements Filter {
         // }, problem.getNbVariables());
         queue = new FibonacciHeap<Variable>(new Key<Variable>() {
             @Override
-            public int getKey(final Variable o1) {
+            public double getKey(final Variable o1) {
                 return o1.getDomainSize();
             }
         }, problem.getNbVariables());
