+++ src/main/java/cspfj/priorityqueues/BinomialHeap.java
@@ -77,8 +77,8 @@ public final class BinomialHeap<T extends Identified> extends AbstractQueue<T> {
             map[id] = node;
         }
 
-        final int oldKey = node.key;
-        final int newKey = key.getKey(arg0);
+        final double oldKey = node.key;
+        final double newKey = key.getKey(arg0);
         node.key = newKey;
 
         if (node.inQueue) {
@@ -129,7 +129,7 @@ public final class BinomialHeap<T extends Identified> extends AbstractQueue<T> {
 
     private int minTree() {
         int min = -1;
-        int minKey = 0;
+        double minKey = 0;
         for (int i = trees.length; --i >= 0;) {
             final BinomialHeapNode<T> tree = trees[i];
             if (tree != null && (min < 0 || tree.key <= minKey)) {
@@ -205,7 +205,7 @@ public final class BinomialHeap<T extends Identified> extends AbstractQueue<T> {
     private void swap(final BinomialHeapNode<T> node1,
             final BinomialHeapNode<T> node2) {
         final T node1Data = node1.data;
-        final int node1Key = node1.key;
+        final double node1Key = node1.key;
         map[node1Data.getId()] = node2;
         map[node2.data.getId()] = node1;
         node1.data = node2.data;
@@ -217,7 +217,7 @@ public final class BinomialHeap<T extends Identified> extends AbstractQueue<T> {
     private static final class BinomialHeapNode<T> {
         private T data;
 
-        private int key;
+        private double key;
 
         private BinomialHeapNode<T> child;
